% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk.R
\name{crosswalk_geoms}
\alias{crosswalk_geoms}
\title{Crosswalk two \code{sf} geoms, using spatial weights.}
\usage{
crosswalk_geoms(
  x_geom,
  y_geom,
  weight_pts,
  weights,
  x_id = NULL,
  y_id = NULL,
  allow_unmatched_weights = c("error", "drop", "distance"),
  verbose = TRUE
)
}
\arguments{
\item{x_geom}{\code{sf} polygons to crosswalk from.}

\item{y_geom}{\code{sf} polygons to crosswalk to.}

\item{weight_pts}{\code{sf} points associated with weights.}

\item{weights}{A vector of weights, same length as \code{weight_pts}.}

\item{x_id}{A vector of ids associated with \code{x_geom}. If \code{NULL}, ids are \code{1:N}.}

\item{y_id}{A vector of ids associated with \code{y_geom}. If \code{NULL}, ids are \code{1:N}.}

\item{allow_unmatched_weights}{How to handle \code{weight_pts} that do not fall in any geom.}

\item{verbose}{Print informative meessages.}
}
\value{
A dataframe with columns \code{x.id}, \code{y.id}, \code{weight}, \code{from_x_to_y}, \code{from_y_to_x}.
If x_id or y_id are NULL, the id columns are integers refering to indices of the geoms.
\code{weight} is the sum of weights in the intersection. \code{from_x_to_y} is the sum of weight in
the intersection divided by the sum of weights in the \code{x_geom} as a whole. It should be
used to proportionally allocate values from \code{x_geom} to \code{y_geom}s.

\code{allow_unmatched_weights} handles unmatched weights. This is useful with Census data,
where boundaries can extend into the water, so block centroids may not fall in a non-Census geom.
If "error", then unmatched weights result in an error. If "drop", unmatched weights are dropped.
If "distance", unmatched weights are assigned to the closest geom.
}
\description{
Crosswalk two \code{sf} geoms, using spatial weights.
}
\examples{
\dontrun{

data("divs_2019")
data("divs_201911")
data("phila_blocks")
cw <- crosswalk_geoms_(
  divs_2019$geometry,
  divs_201911$geometry,
  weight_pts=phila_blocks$geometry,
  weights=phila_blocks$pop,
  x_id=divs_2019$warddiv,
  y_id=divs_201911$warddiv
)
}
}
