% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk_area.R
\name{crosswalk_geoms_area}
\alias{crosswalk_geoms_area}
\title{Crosswalk two \code{sf} geoms, using area.}
\usage{
crosswalk_geoms_area(
  x_geom,
  y_geom,
  x_id = NULL,
  y_id = NULL,
  allow_unmatched_weights = c("error", "allow"),
  verbose = TRUE,
  tol = 1e-04
)
}
\arguments{
\item{x_geom}{\code{sf} polygons to crosswalk from.}

\item{y_geom}{\code{sf} polygons to crosswalk to.}

\item{x_id}{A vector of ids associated with \code{x_geom}. If \code{NULL}, ids are \code{1:N}.}

\item{y_id}{A vector of ids associated with \code{y_geom}. If \code{NULL}, ids are \code{1:N}.}

\item{allow_unmatched_weights}{How to handle cases where geoms don't perfectly overlap.}

\item{verbose}{Print informative meessages.}

\item{tol}{A tolerance used if \code{allow_unmatched_weights} is "error".}
}
\value{
A dataframe with columns \code{x.id}, \code{y.id}, \code{area}, \code{from_x_to_y}, \code{from_y_to_x}.
If x_id or y_id are NULL, the id columns are integers refering to indices of the geoms.
\code{area} is the area of the intersection. \code{from_x_to_y} is the area in
the intersection divided by the area of \code{x_geom} as a whole. It should be
used to proportionally allocate values from \code{x_geom} to \code{y_geom}s.

\code{allow_unmatched_weights} handles unmatched geoms that are not 1:1. If "error",
then an error is raised whenever the total area for a given geom in the final crosswalk
divided by the original area of the geom is not within \code{tol} of 1.
}
\description{
Crosswalk two \code{sf} geoms, using area.
}
\examples{
\dontrun{

data("divs_2019")
data("divs_201911")
cw <- crosswalk_geoms_area(
  divs_2019$geometry,
  divs_201911$geometry,
  x_id=divs_2019$warddiv,
  y_id=divs_201911$warddiv
)
}
}
